+++ Assets/_Application/_Twin.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8bb47baeeb5df449b81303d545eb3cbf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
